snd: set period bytes to 0xc000
Was previously set to 0x18000 because guess-and-check work made that
work with OBS. Set it to 0xc000 because the GV-USB2 generally sends 0xc0
bytes per isoc packet and we tell it to give us 0x100 isoc packets worth
of data each time. This multiplies to 0xc000 per data-exchange on
average. I believe OBS didn't like data coming in at non-regular
intervals and threw a sound-tantrum as a result. 0x18000 worked because
it was a multiple of 0xc000.