gvusb2-vid: add capture start offset controls
Added two custom controls, Vertical Start and Horizontal Start. Each
controls where in the video signal the capture card starts capturing
from. In addition to this, the pixel format was changed from YUYV to
the proper UYVY. The Windows driver actually spits out YUYV for some
reason, but this is because it uses a horizontal capture start that
isn't a multiple of four. Thus capture started at the 3rd byte and
shifted UYVY to YUYV.