gvusb2-snd: fix an oops by not killing freed urbs
If the GV-USB2 is removed while a program is accessing the alsa end,
urbs get released and as a result, when the alsa end is released and the
driver tries to kill the released urbs, the driver oopses. This is fixed
by freeing and nulling out the urbs on disconnect so that when killing
urbs, it won't kill ones that are already nulled out.